requirements.txt
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.28.0
pillow>=9.5.0
scikit-learn>=1.3.0
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
jupyter>=1.0.0
